Refactored estimators as preparation to replacing tvrage as the main series estimator. Created a new 'estimate_release_series' plugin group and added tvrage and internal db estimators (split from main file) to it.
